Rewrite assembly to not need codegen
The original assembly relied on a build-time generated "gen_defines.asm" that contains various
values and struct offsets needed by the assembly. This creates a lot of friction, as this build step
must account for the numerous build environments and configurations that are possible, on top
of needing an alternative path to handle cross-compilation. By modifying the assembly to no use
extern variables, asm_offsets.c can contain all the necessary values as a compile-time value, such
that at link time the values are available to the assembler. This does slightly increase runtime as
we are replacing constants with variables, but the added benefits of simplifying the build step
outweighs it.